hole_puncher - Extracted hole puncher hooks
