+++ hole_puncher/src/hooks/state.rs
@@ -1,0 +1,12 @@
+pub struct State {
+    // All known peers
+    pub peers: Vec<String>
+}
+
+impl State {
+    pub fn new() -> Self {
+        Self {
+            peers: Vec::new()
+        }
+    }
+}
